Add modules for reading and saving trees, labelling of leaves
read_tree.py takes in a .pkl file and prints out the tree.
save_tree.py generates a random tree and saves into .pkl file.
Also added labelling of leaves and fixed a bug which partitions leaves into two bins.
Next step would be to use the same tree for the ai_playground.py so we can actually
do the scoring function to check for misassignments.